LayoutTests/media/video-concurrent-visible-playback-expected.txt


Test autoplay when both invisible playback and concurrent playback are not allowed.

RUN(video.src = findMediaFile("video", "content/test"))
RUN(video.src = findMediaFile("video", "content/test"))
RUN(video.src = findMediaFile("video", "content/test"))

EVENT(playing)
EXPECTED (videos[0].paused == 'true') OK
EXPECTED (videos[1].paused == 'false') OK
EXPECTED (videos[2].paused == 'true') OK

END OF TEST
